Refactor execute command module layout
