--- crates/perl-lsp/src/execute_command.rs
@@ -1,1943 +1,0 @@
-//! Execute command support for running tests and debugging.
-//!
-//! This module provides comprehensive support for the LSP executeCommand request,
-//! enabling seamless integration between editors and Perl development workflows.
-//! It implements the dual analyzer strategy for code quality analysis with 100% availability.
-//!
-//! ## LSP Workflow Integration
-//!
-//! The executeCommand implementation follows the Parse → Index → Navigate → Complete → Analyze workflow:
-//! - **Parse**: Source files are parsed using the perl-parser for syntax validation
-//! - **Index**: Command metadata is indexed for efficient command resolution
-//! - **Navigate**: Commands provide navigation to test results and diagnostic locations
-//! - **Complete**: Auto-completion for command parameters and test subroutines
-//! - **Analyze**: Comprehensive code quality analysis via dual analyzer strategy
-//!
-//! ## Performance Characteristics
-//!
-//! - **Command execution**: <50ms response time for code actions
-//! - **executeCommand processing**: <2s execution time for comprehensive analysis
-//! - **Memory usage**: <10MB for typical Perl file analysis
-//! - **Incremental analysis**: Leverages ≤1ms parsing SLO for real-time feedback
-//!
-//! ## Supported Commands
-//!
-//! ```no_run
-//! use perl_lsp::execute_command::{ExecuteCommandProvider, get_supported_commands};
-//! use serde_json::Value;
-//!
-//! let provider = ExecuteCommandProvider::new();
-//! let commands = get_supported_commands();
-//!
-//! // Execute perl.runCritic command with dual analyzer strategy
-//! let result = provider.execute_command(
-//!     "perl.runCritic",
-//!     vec![Value::String("/path/to/file.pl".to_string())]
-//! );
-//! ```
-//!
-//! ## Error Recovery
-//!
-//! Commands implement comprehensive error recovery strategies:
-//! - **File not found**: Graceful error responses with actionable feedback
-//! - **Syntax errors**: Parse error detection with location information
-//! - **External tool failures**: Automatic fallback to built-in analyzers
-//! - **Permission errors**: Clear error messages with resolution suggestions
-
-//!   Execute command implementation for Perl LSP with dual analyzer strategy.
-//!
-//! This module provides comprehensive executeCommand support for the Perl Language Server,
-//! implementing a dual analyzer strategy that combines external tool integration with
-//! built-in fallback analysis. The implementation ensures 100% availability and robust
-//! security through workspace root enforcement and path traversal protection.
-//!
-//! # Architecture
-//!
-//! The module follows a dual analyzer pattern:
-//! - **External Tools**: Integrates with perlcritic, perltidy, and test runners
-//! - **Built-in Fallback**: Provides analysis when external tools are unavailable
-//! - **Security-First**: All file operations are workspace-enforced with canonicalization
-//! - **LSP Compliant**: Proper JSON-RPC error handling and timeout management
-//!
-//! # Supported Commands
-//!
-//! - `perl.runCritic` - Code quality analysis with dual analyzer strategy
-//! - `perl.runFile` - Execute Perl scripts with structured output
-//! - `perl.runTest` - Execute a single discovered test case
-//! - `perl.runTestFile` - Execute a specific test file directly
-//! - `perl.runTests` - Run test suites with coverage reporting
-//! - `perl.runTestSub` - Execute individual test subroutines
-//! - `perl.debugFile` - Launch file-level debugging
-//!
-//! # Examples
-//!
-//! ```no_run
-//! use perl_lsp::execute_command::{ExecuteCommandProvider, command_exists};
-//! use serde_json::Value;
-//!
-//! // Create provider with workspace security
-//! let provider = ExecuteCommandProvider::with_workspace_roots(
-//!     vec!["/home/user/project".into()]
-//! );
-//!
-//! // Execute command with secure path resolution
-//! let result = provider.execute_command(
-//!     "perl.runCritic",
-//!     vec![Value::String("file:///home/user/project/script.pl".to_string())]
-//! );
-//!
-//! // Check tool availability
-//! if command_exists("perlcritic") {
-//!     println!("External perlcritic available");
-//! }
-//! ```
-
-use crate::perl_critic::{BuiltInAnalyzer, CriticAnalyzer, CriticConfig};
-use crate::protocol::JsonRpcError;
-use serde::{Deserialize, Serialize};
-use serde_json::{Value, json};
-use std::path::{Path, PathBuf};
-use std::process::Command;
-
-// Cross-platform helpers for synthesizing `ExitStatus` in tests/mocks.
-#[cfg(any(test, doctest))]
-#[cfg(unix)]
-use std::os::unix::process::ExitStatusExt as _;
-#[cfg(any(test, doctest))]
-#[cfg(windows)]
-use std::os::windows::process::ExitStatusExt as _;
-
-// Map a logical exit code (0/1/…) to the platform's raw representation.
-#[cfg(any(test, doctest))]
-#[cfg(unix)]
-#[inline]
-fn raw_exit(code: i32) -> i32 {
-    // POSIX: wait(2) encodes exit code in the high byte.
-    code << 8
-}
-#[cfg(any(test, doctest))]
-#[cfg(windows)]
-#[inline]
-fn raw_exit(code: i32) -> u32 {
-    // Windows: raw is the process' exit code directly.
-    code as u32
-}
-
-// Future platforms: fail fast during tests so we notice and add a mapping.
-// Only enforced for test/doctest builds to avoid breaking non-Unix/Windows release targets.
-#[cfg(all(any(test, doctest), not(any(unix, windows))))]
-compile_error!("Add raw_exit() mapping for this platform.");
-
-// Helper to reduce duplication in tests while keeping the trait requirement localized.
-#[cfg(any(test, doctest))]
-#[inline]
-fn mock_status(code: i32) -> std::process::ExitStatus {
-    std::process::ExitStatus::from_raw(raw_exit(code))
-}
-
-/// Commands supported by the Perl LSP server for test execution and code analysis.
-///
-/// This enum defines all supported executeCommand requests that can be invoked from
-/// LSP-compatible editors. Each command provides specific functionality for Perl
-/// development workflows with comprehensive error handling and result formatting.
-///
-/// # Examples
-///
-/// ```no_run
-/// use perl_lsp::execute_command::PerlCommand;
-/// use serde_json;
-///
-/// // Deserialize command from LSP request
-/// let json = r#"{"runTests": {"filePath": "/path/to/test.pl"}}"#;
-/// let command: Result<PerlCommand, _> = serde_json::from_str(json);
-/// ```
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum PerlCommand {
-    /// Run all tests in a file using appropriate test runner (prove or perl).
-    ///
-    /// Automatically detects test files (.t extension, /t/ directory, or 'test' in name)
-    /// and uses the optimal execution strategy for maximum compatibility.
-    RunTests {
-        /// Path to the Perl test file to execute
-        file_path: String,
-    },
-    /// Run a specific test subroutine with enhanced error detection.
-    ///
-    /// Executes a named subroutine within a test file, providing targeted
-    /// test execution for faster development feedback cycles.
-    RunTestSub {
-        /// Path to the Perl file containing the subroutine
-        file_path: String,
-        /// Name of the subroutine to execute
-        sub_name: String,
-    },
-    /// Run a Perl file directly with the perl interpreter.
-    ///
-    /// Provides direct execution of Perl scripts with standardized result formatting
-    /// and comprehensive error capture for development workflows.
-    RunFile {
-        /// Path to the Perl file to execute
-        file_path: String,
-    },
-    /// Debug a test file (placeholder for future DAP integration).
-    ///
-    /// Reserved for future Debug Adapter Protocol integration. Currently returns
-    /// a structured response indicating debugging support is planned.
-    DebugTests {
-        /// Path to the test file for debugging
-        file_path: String,
-    },
-}
-
-/// Result of executing a command with standardized structure.
-///
-/// All executeCommand operations return results in this consistent format,
-/// enabling reliable error handling and result processing in LSP clients.
-///
-/// # Examples
-///
-/// ```
-/// use perl_lsp::execute_command::CommandResult;
-///
-/// let result = CommandResult {
-///     success: true,
-///     output: "Tests passed successfully".to_string(),
-///     error: None,
-/// };
-/// ```
-#[derive(Debug, Serialize)]
-pub struct CommandResult {
-    /// Whether the command executed successfully
-    pub success: bool,
-    /// Standard output from the command execution
-    pub output: String,
-    /// Error message if the command failed, None if successful
-    pub error: Option<String>,
-}
-
-/// Execute command provider implementing the LSP executeCommand method.
-///
-/// This provider handles all supported Perl LSP commands with comprehensive error
-/// handling, dual analyzer strategy for code quality, and performance optimization.
-/// It integrates seamlessly with the LSP workflow for enterprise-grade functionality.
-///
-/// # Performance
-///
-/// - Command resolution: <1ms using efficient routing
-/// - Code analysis: <2s for comprehensive quality checks
-/// - Memory usage: <10MB for typical Perl files
-/// - Thread safety: Fully thread-safe for concurrent LSP requests
-///
-/// # Examples
-///
-/// ```no_run
-/// use perl_lsp::execute_command::ExecuteCommandProvider;
-/// use serde_json::Value;
-///
-/// let provider = ExecuteCommandProvider::new();
-///
-/// // Execute code quality analysis
-/// let result = provider.execute_command(
-///     "perl.runCritic",
-///     vec![Value::String("/path/to/file.pl".to_string())]
-/// );
-///
-/// match result {
-///     Ok(response) => println!("Analysis completed: {:?}", response),
-///     Err(error) => eprintln!("Command failed: {}", error),
-/// }
-/// ```
-pub struct ExecuteCommandProvider {
-    /// Workspace root paths for security enforcement
-    workspace_roots: Vec<PathBuf>,
-}
-
-impl Default for ExecuteCommandProvider {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl ExecuteCommandProvider {
-    /// Create a new execute command provider.
-    ///
-    /// Initializes the provider with default configuration optimized for
-    /// performance and reliability in LSP environments.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use perl_lsp::execute_command::ExecuteCommandProvider;
-    ///
-    /// let provider = ExecuteCommandProvider::new();
-    /// ```
-    pub fn new() -> Self {
-        Self { workspace_roots: Vec::new() }
-    }
-
-    /// Create a new execute command provider with workspace root enforcement.
-    ///
-    /// This constructor enables path traversal protection by enforcing that all
-    /// file operations must be within the specified workspace root directories.
-    ///
-    /// # Arguments
-    ///
-    /// * `workspace_roots` - The root directory paths to enforce for security
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use perl_lsp::execute_command::ExecuteCommandProvider;
-    /// use std::path::PathBuf;
-    ///
-    /// let provider = ExecuteCommandProvider::with_workspace_roots(
-    ///     vec![PathBuf::from("/home/user/project")]
-    /// );
-    /// ```
-    pub fn with_workspace_roots(workspace_roots: Vec<PathBuf>) -> Self {
-        Self { workspace_roots }
-    }
-
-    /// Execute a command with comprehensive error handling and argument validation.
-    ///
-    /// This is the main entry point for LSP executeCommand requests. It provides
-    /// routing to specific command implementations with consistent error handling
-    /// and response formatting.
-    ///
-    /// # Arguments
-    ///
-    /// * `command` - The command identifier (e.g., "perl.runCritic")
-    /// * `arguments` - Command arguments as JSON values
-    ///
-    /// # Returns
-    ///
-    /// Returns a JSON response with standardized structure or an error message.
-    /// All successful responses include status, output, and metadata fields.
-    ///
-    /// # Performance
-    ///
-    /// - Command routing: <1ms for all supported commands
-    /// - Argument validation: <1ms with comprehensive type checking
-    /// - Total overhead: <2ms excluding actual command execution
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use perl_lsp::execute_command::ExecuteCommandProvider;
-    /// use serde_json::Value;
-    ///
-    /// let provider = ExecuteCommandProvider::new();
-    ///
-    /// // Run code quality analysis
-    /// let result = provider.execute_command(
-    ///     "perl.runCritic",
-    ///     vec![Value::String("/path/to/file.pl".to_string())]
-    /// );
-    ///
-    /// // Run specific test subroutine
-    /// let test_result = provider.execute_command(
-    ///     "perl.runTestSub",
-    ///     vec![
-    ///         Value::String("/path/to/test.pl".to_string()),
-    ///         Value::String("test_function".to_string())
-    ///     ]
-    /// );
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns `Err(String)` for:
-    /// - Unknown command identifiers
-    /// - Missing or invalid arguments
-    /// - File access errors
-    /// - Command execution failures
-    pub fn execute_command(&self, command: &str, arguments: Vec<Value>) -> Result<Value, String> {
-        match command {
-            "perl.runTests" => {
-                let file_path = self.resolve_path_from_args(&arguments)?;
-                self.run_tests(&file_path)
-            }
-            "perl.runFile" => {
-                let file_path = self.resolve_path_from_args(&arguments)?;
-                self.run_file(&file_path)
-            }
-            "perl.runTestSub" => {
-                let file_path = self.resolve_path_from_args(&arguments)?;
-                let sub_name = arguments
-                    .get(1)
-                    .and_then(|v| v.as_str())
-                    .ok_or_else(|| "Missing subroutine name argument".to_string())?;
-                self.run_test_sub(&file_path, sub_name)
-            }
-            "perl.debugTests" => {
-                let file_path = self.resolve_path_from_args(&arguments)?;
-                self.debug_tests(&file_path)
-            }
-            "perl.runCritic" => {
-                // Use secure path resolution instead of extract_file_path_argument
-                self.run_critic_secure(&arguments)
-            }
-            _ => Err(format!("Unknown command: {}", command)),
-        }
-    }
-
-    /// Run all tests in a file using appropriate test runner
-    fn run_tests(&self, file_path: &Path) -> Result<Value, String> {
-        let file_path_str = file_path.to_string_lossy();
-        let is_test_file = self.is_test_file(&file_path_str);
-        let (command_name, mut cmd) = if is_test_file && self.command_exists("prove") {
-            ("prove", {
-                let mut cmd = Command::new("prove");
-                cmd.arg("-v").arg("--").arg(file_path.as_os_str());
-                cmd
-            })
-        } else {
-            ("perl", {
-                let mut cmd = Command::new("perl");
-                cmd.arg("--").arg(file_path.as_os_str());
-                cmd
-            })
-        };
-
-        let result = cmd.output().map_err(|e| format!("Failed to run {}: {}", command_name, e))?;
-
-        Ok(self.format_command_result(result, Some(("command", command_name.into()))))
-    }
-
-    /// Run a specific test subroutine with enhanced error handling
-    fn run_test_sub(&self, file_path: &Path, sub_name: &str) -> Result<Value, String> {
-        // Enhanced subroutine invocation with better error detection
-        // Use @ARGV to safely pass file path and subroutine name preventing code injection
-        let perl_code = r#"
-            my ($file, $sub) = @ARGV;
-            do $file;
-            if (defined &$sub) {
-                no strict 'refs';
-                &$sub();
-            } else {
-                die "Subroutine $sub not found";
-            }
-        "#;
-
-        let result = Command::new("perl")
-            .arg("-e")
-            .arg(perl_code)
-            .arg("--")
-            .arg(file_path.as_os_str())
-            .arg(sub_name)
-            .output()
-            .map_err(|e| format!("Failed to run test subroutine: {}", e))?;
-
-        Ok(self.format_command_result(result, Some(("subroutine", sub_name.into()))))
-    }
-
-    /// Run a Perl file with standardized result formatting
-    fn run_file(&self, file_path: &Path) -> Result<Value, String> {
-        let result = Command::new("perl")
-            .arg("--") // Safety against argument injection
-            .arg(file_path.as_os_str())
-            .output()
-            .map_err(|e| format!("Failed to run file: {}", e))?;
-
-        Ok(self.format_command_result(result, None))
-    }
-
-    /// Debug tests (placeholder for future implementation)
-    fn debug_tests(&self, file_path: &Path) -> Result<Value, String> {
-        // For now, just run with perl -d
-        // In the future, this could integrate with Perl debugger or DAP
-        let file_path_str = file_path.to_string_lossy();
-        Ok(json!({
-            "success": false,
-            "output": format!("Debug mode not yet implemented for {}", file_path_str),
-            "error": Some("Debugging support coming soon".to_string())
-        }))
-    }
-
-    /// Run Perl::Critic analysis using dual analyzer strategy with secure path resolution
-    fn run_critic_secure(&self, arguments: &[Value]) -> Result<Value, String> {
-        // Use secure path resolution with workspace enforcement
-        let canonical_path = match self.resolve_path_from_args(arguments) {
-            Ok(path) => path,
-            Err(e) => {
-                // Missing arguments are validation errors - must fail with Err
-                if e.contains("Missing file path argument") {
-                    return Err(e);
-                }
-
-                // Handle file not found errors gracefully with structured error response
-                // IMPORTANT: Preserve the path in the error message for debugging
-                if e.contains("File not found")
-                    || e.contains("does not exist")
-                    || e.contains("No such file or directory")
-                    || e.contains("Failed to canonicalize")
-                {
-                    // Extract and preserve the full error message which includes the path
-                    let error_message = if e.contains("Failed to canonicalize") {
-                        // Extract path from "Failed to canonicalize path 'X': Y"
-                        if let Some(start) = e.find("'") {
-                            if let Some(end) = e[start + 1..].find("'") {
-                                let path = &e[start + 1..start + 1 + end];
-                                format!("File not found: {}", path)
-                            } else {
-                                "File not found".to_string()
-                            }
-                        } else {
-                            "File not found".to_string()
-                        }
-                    } else {
-                        // For "File not found: X" errors, preserve as-is
-                        e.clone()
-                    };
-                    return Ok(self.format_critic_error(error_message, "none"));
-                }
-
-                // Security-related errors (workspace traversal, length, ..) are failures
-                if e.contains("Path traversal")
-                    || e.contains("outside workspace")
-                    || e.contains("Argument too long")
-                {
-                    return Err(format!("Path resolution failed: {}", e));
-                }
-
-                // All other errors are handled gracefully
-                return Ok(self.format_critic_error(e, "none"));
-            }
-        };
-
-        // Dual analyzer strategy: external perlcritic with built-in fallback
-        if command_exists("perlcritic") {
-            match self.run_external_critic(&canonical_path) {
-                Ok(result) => return Ok(result),
-                Err(_) => {
-                    // Silently fall back to built-in analyzer for seamless UX
-                }
-            }
-        }
-
-        // Built-in analyzer provides 100% availability
-        self.run_builtin_critic(&canonical_path)
-    }
-
-    /// Run Perl::Critic analysis using dual analyzer strategy (legacy method - deprecated)
-    ///
-    /// # Security Warning
-    ///
-    /// This method is deprecated and vulnerable to path traversal attacks.
-    /// Use `run_critic_secure` instead for secure path resolution.
-    #[deprecated(since = "0.8.9", note = "Use run_critic_secure for secure path resolution")]
-    #[allow(dead_code)]
-    #[allow(deprecated)]
-    fn run_critic(&self, file_path: &str) -> Result<Value, String> {
-        let normalized_path = self.normalize_file_path(file_path);
-        let path = Path::new(normalized_path);
-
-        if !path.exists() {
-            return Ok(
-                self.format_critic_error(format!("File not found: {}", normalized_path), "none")
-            );
-        }
-
-        // Dual analyzer strategy: external perlcritic with built-in fallback
-        if command_exists("perlcritic") {
-            match self.run_external_critic(path) {
-                Ok(result) => return Ok(result),
-                Err(_) => {
-                    // Silently fall back to built-in analyzer for seamless UX
-                }
-            }
-        }
-
-        // Built-in analyzer provides 100% availability
-        self.run_builtin_critic(path)
-    }
-
-    /// Run external perlcritic with standardized response formatting
-    fn run_external_critic(&self, file_path: &Path) -> Result<Value, String> {
-        let config = CriticConfig {
-            severity: 3, // Harsh and above for production-quality analysis
-            verbose: true,
-            ..Default::default()
-        };
-
-        let mut analyzer = CriticAnalyzer::with_os_runtime(config);
-
-        match analyzer.analyze_file(file_path) {
-            Ok(violations) => {
-                let formatted_violations: Vec<_> = violations
-                    .iter()
-                    .map(|v| {
-                        self.format_violation(
-                            &v.policy,
-                            &v.description,
-                            &v.explanation,
-                            v.severity as u8,
-                            (v.range.start.line + 1) as usize,
-                            (v.range.start.column + 1) as usize,
-                            &v.file,
-                        )
-                    })
-                    .collect();
-
-                Ok(json!({
-                    "status": "success",
-                    "violations": formatted_violations,
-                    "violationCount": formatted_violations.len(),
-                    "analyzerUsed": "external"
-                }))
-            }
-            Err(e) => Err(format!("External perlcritic failed: {}", e)),
-        }
-    }
-
-    /// Run built-in critic analyzer with comprehensive error handling
-    fn run_builtin_critic(&self, file_path: &Path) -> Result<Value, String> {
-        use crate::Parser;
-
-        let content = std::fs::read_to_string(file_path)
-            .map_err(|e| format!("Failed to read file: {}", e))?;
-
-        let code_text = perl_parser::util::code_slice(&content);
-        let mut parser = Parser::new(code_text);
-
-        let (ast, parse_error) = match parser.parse() {
-            Ok(ast) => (ast, None),
-            Err(error) => {
-                let message = error.to_string();
-                (
-                    crate::ast::Node::new(
-                        crate::ast::NodeKind::Error {
-                            message,
-                            expected: vec![],
-                            found: None,
-                            partial: None,
-                        },
-                        crate::ast::SourceLocation { start: 0, end: code_text.len() },
-                    ),
-                    Some(error),
-                )
-            }
-        };
-
-        let analyzer = BuiltInAnalyzer::new();
-        let mut all_violations = analyzer.analyze(&ast, code_text);
-        if let Some(error) = parse_error {
-            all_violations.push(self.create_syntax_error_violation(&error, code_text, file_path));
-        }
-
-        let formatted_violations: Vec<_> = all_violations
-            .iter()
-            .map(|v| {
-                self.format_violation(
-                    &v.policy,
-                    &v.description,
-                    &v.explanation,
-                    v.severity as u8,
-                    (v.range.start.line + 1) as usize,
-                    (v.range.start.column + 1) as usize,
-                    &file_path.to_string_lossy(),
-                )
-            })
-            .collect();
-
-        Ok(json!({
-            "status": "success",
-            "violations": formatted_violations,
-            "violationCount": formatted_violations.len(),
-            "analyzerUsed": "builtin"
-        }))
-    }
-
-    /// Check if a file path appears to be a test file
-    fn is_test_file(&self, file_path: &str) -> bool {
-        file_path.ends_with(".t") || file_path.contains("/t/") || file_path.contains("test")
-    }
-
-    /// Format command execution result with consistent structure
-    fn format_command_result(
-        &self,
-        result: std::process::Output,
-        extra_field: Option<(&str, Value)>,
-    ) -> Value {
-        let output = String::from_utf8_lossy(&result.stdout);
-        let error = if !result.status.success() {
-            Some(String::from_utf8_lossy(&result.stderr).to_string())
-        } else {
-            None
-        };
-
-        let mut response = json!({
-            "success": result.status.success(),
-            "output": output.to_string(),
-            "error": error
-        });
-
-        if let Some((key, value)) = extra_field {
-            response[key] = value;
-        }
-
-        response
-    }
-
-    /// Securely resolve a file path from command arguments with workspace root enforcement.
-    ///
-    /// This method provides comprehensive path traversal protection by:
-    /// - Normalizing file:// URIs to local file paths
-    /// - Canonicalizing paths to resolve .. and . components
-    /// - Enforcing workspace root boundaries when configured
-    /// - Validating file existence and readability
-    ///
-    /// # Arguments
-    ///
-    /// * `arguments` - Command arguments containing the file path
-    ///
-    /// # Returns
-    ///
-    /// A canonicalized `PathBuf` if the path is valid and within workspace bounds
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - No file path argument is provided
-    /// - Path contains invalid characters or traversal attempts
-    /// - Path is outside the workspace root (if configured)
-    /// - File does not exist or is not readable
-    ///
-    /// # Security
-    ///
-    /// This method prevents path traversal attacks by canonicalizing paths
-    /// and enforcing workspace boundaries. All paths are resolved relative
-    /// to the workspace root when configured.
-    fn resolve_path_from_args(&self, arguments: &[Value]) -> Result<PathBuf, String> {
-        // Extract the file path argument
-        let raw_path = arguments
-            .first()
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "Missing file path argument".to_string())?;
-
-        // Defense in depth: cap argument length to prevent abuse
-        const MAX_ARG_LENGTH: usize = 4096;
-        if raw_path.len() > MAX_ARG_LENGTH {
-            return Err(format!(
-                "Argument too long ({} bytes, max {})",
-                raw_path.len(),
-                MAX_ARG_LENGTH
-            ));
-        }
-
-        // Normalize file:// URIs
-        let normalized_path = raw_path.strip_prefix("file://").unwrap_or(raw_path);
-
-        // Defense in depth: reject paths with parent traversal components
-        // even though canonicalize() resolves them, this catches attempts early
-        if normalized_path.contains("..") {
-            return Err("Path traversal attempt detected: path contains '..' component".to_string());
-        }
-
-        // Convert to PathBuf and canonicalize to resolve .. and . components
-        let path = Path::new(normalized_path);
-        let canonical_path = path
-            .canonicalize()
-            .map_err(|e| format!("Failed to canonicalize path '{}': {}", normalized_path, e))?;
-
-        // Determine workspace boundaries
-        // Security: When workspace_roots is empty (single-file mode), use CWD as the
-        // fallback boundary to prevent unrestricted path traversal. This ensures that
-        // even without explicit workspace configuration, files outside the working
-        // directory cannot be accessed via executeCommand.
-        let effective_roots: Vec<PathBuf> = if self.workspace_roots.is_empty() {
-            // Fallback: use CWD as boundary when no workspace roots configured
-            // This prevents unrestricted path traversal in single-file mode
-            match std::env::current_dir() {
-                Ok(cwd) => vec![cwd],
-                Err(_) => {
-                    return Err(
-                        "No workspace roots configured and cannot determine working directory"
-                            .to_string(),
-                    );
-                }
-            }
-        } else {
-            self.workspace_roots.clone()
-        };
-
-        let mut allowed = false;
-        for workspace_root in &effective_roots {
-            if let Ok(canonical_root) = workspace_root.canonicalize() {
-                if canonical_path.starts_with(&canonical_root) {
-                    allowed = true;
-                    break;
-                }
-            }
-        }
-
-        if !allowed {
-            return Err(format!(
-                "Path traversal detected: {} is outside workspace boundaries",
-                canonical_path.display()
-            ));
-        }
-
-        // Validate file existence and readability
-        if !canonical_path.exists() {
-            return Err(format!("File not found: {}", canonical_path.display()));
-        }
-
-        if !canonical_path.is_file() {
-            return Err(format!("Path is not a file: {}", canonical_path.display()));
-        }
-
-        // Check basic readability (this will fail fast if permissions are wrong)
-        std::fs::metadata(&canonical_path).map_err(|e| {
-            format!("Cannot read file metadata '{}': {}", canonical_path.display(), e)
-        })?;
-
-        Ok(canonical_path)
-    }
-
-    /// Resolve a debug file path with the same workspace security as other commands.
-    ///
-    /// Wraps `resolve_path_from_args` for a single string path argument,
-    /// providing the same path traversal protection and workspace enforcement.
-    pub fn resolve_debug_file_path(&self, file_path: &str) -> Result<PathBuf, String> {
-        self.resolve_path_from_args(&[Value::String(file_path.to_string())])
-    }
-
-    /// Normalize file path by handling URI schemes and path formats (legacy method - deprecated)
-    ///
-    /// # Security Warning
-    ///
-    /// This method is deprecated and vulnerable to path traversal attacks.
-    /// Use `resolve_path_from_args` instead for secure path resolution.
-    #[deprecated(since = "0.8.9", note = "Use resolve_path_from_args for secure path resolution")]
-    #[allow(dead_code)]
-    fn normalize_file_path<'a>(&self, file_path: &'a str) -> &'a str {
-        file_path.strip_prefix("file://").unwrap_or(file_path)
-    }
-
-    /// Format a violation with consistent structure
-    fn format_violation(
-        &self,
-        policy: &str,
-        description: &str,
-        explanation: &str,
-        severity: u8,
-        line: usize,
-        column: usize,
-        file: &str,
-    ) -> Value {
-        json!({
-            "policy": policy,
-            "description": description,
-            "explanation": explanation,
-            "severity": severity,
-            "line": line,
-            "column": column,
-            "file": file
-        })
-    }
-
-    /// Format critic error response with consistent structure
-    fn format_critic_error(&self, error_message: String, analyzer_used: &str) -> Value {
-        json!({
-            "status": "error",
-            "error": error_message,
-            "violations": [],
-            "violationCount": 0,
-            "analyzerUsed": analyzer_used
-        })
-    }
-
-    /// Create a syntax error violation from parse error
-    fn create_syntax_error_violation(
-        &self,
-        error: &perl_parser::ParseError,
-        _content: &str,
-        file_path: &Path,
-    ) -> crate::perl_critic::Violation {
-        let error_msg = format!("{}", error);
-        let (line, column) = (0, 0); // Default for parse errors
-
-        crate::perl_critic::Violation {
-            policy: "Syntax::ParseError".to_string(),
-            description: format!("Syntax error: {}", error_msg),
-            explanation: "This code contains a syntax error that prevents parsing. Fix the syntax error before running additional checks.".to_string(),
-            severity: crate::perl_critic::Severity::Brutal, // Critical severity for syntax errors
-            range: crate::position::Range {
-                start: crate::position::Position { byte: 0, line: line as u32, column: column as u32 },
-                end: crate::position::Position { byte: 1, line: line as u32, column: (column + 1) as u32 },
-            },
-            file: file_path.to_string_lossy().to_string(),
-        }
-    }
-
-    /// Check if a command exists in PATH with cross-platform compatibility
-    fn command_exists(&self, command: &str) -> bool {
-        Command::new("which")
-            .arg(command)
-            .output()
-            .map(|output| output.status.success())
-            .unwrap_or(false)
-    }
-}
-
-/// Check if a command exists in the system PATH without using external tools.
-///
-/// This function provides a portable way to check command availability by
-/// attempting to execute the command with `--version` flag and checking
-/// if it succeeds. This avoids dependency on `which` or similar utilities.
-///
-/// # Arguments
-///
-/// * `command` - The command name to check
-///
-/// # Returns
-///
-/// `true` if the command exists and is executable, `false` otherwise
-///
-/// # Examples
-///
-/// ```no_run
-/// use perl_lsp::execute_command::command_exists;
-///
-/// if command_exists("perlcritic") {
-///     println!("perlcritic is available");
-/// }
-/// ```
-pub fn command_exists(command: &str) -> bool {
-    std::process::Command::new(command)
-        .arg("--version")
-        .output()
-        .map(|output| output.status.success())
-        .unwrap_or(false)
-}
-
-/// Get the list of supported commands for LSP executeCommand capability.
-///
-/// Returns all command identifiers that can be executed via the LSP executeCommand
-/// method. This list is used for capability registration and command validation.
-///
-/// # Returns
-///
-/// A vector of command identifiers including:
-/// - `perl.runTests`: Execute all tests in a file
-/// - `perl.runFile`: Run a Perl file directly
-/// - `perl.runTestSub`: Execute a specific test subroutine
-/// - `perl.runCritic`: Perform code quality analysis
-/// - `perl.runTest`: Run a single test
-/// - `perl.runTestFile`: Run a test file
-/// - `perl.debugFile`: Debug a Perl file
-///
-/// # Examples
-///
-/// ```
-/// use perl_lsp::execute_command::get_supported_commands;
-///
-/// let commands = get_supported_commands();
-/// assert!(commands.contains(&"perl.runCritic".to_string()));
-/// assert_eq!(commands.len(), 7);
-/// ```
-///
-/// # Performance
-///
-/// - Execution time: <1ms (static list generation)
-/// - Memory usage: <1KB for command list
-pub fn get_supported_commands() -> Vec<String> {
-    crate::protocol::capabilities::get_supported_commands()
-}
-
-/// Command executor for LSP incremental server with proper JSON-RPC error handling.
-///
-/// This struct provides a bridge between the incremental LSP server and the
-/// ExecuteCommandProvider, ensuring that errors are returned as proper JSON-RPC
-/// errors rather than embedded in the result payload.
-///
-/// # JSON-RPC Error Mapping
-///
-/// - Invalid arguments → `-32602` (InvalidParams)
-/// - Unknown commands → `-32601` (MethodNotFound)
-/// - Security violations → `-32603` (InternalError)
-/// - General failures → `-32603` (InternalError)
-///
-/// # Examples
-///
-/// ```no_run
-/// use perl_lsp::execute_command::CommandExecutor;
-/// use serde_json::Value;
-///
-/// let executor = CommandExecutor::new();
-/// let result = executor.execute("perl.runCritic", Some(&vec![
-///     Value::String("file:///path/to/file.pl".to_string())
-/// ]));
-/// ```
-pub struct CommandExecutor {
-    provider: ExecuteCommandProvider,
-}
-
-impl Default for CommandExecutor {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl CommandExecutor {
-    /// Create a new command executor with default configuration.
-    ///
-    /// The executor is initialized with workspace-agnostic configuration.
-    /// For workspace-aware security enforcement, use `with_workspace_roots`.
-    pub fn new() -> Self {
-        Self { provider: ExecuteCommandProvider::new() }
-    }
-
-    /// Create a new command executor with workspace root enforcement.
-    ///
-    /// This constructor enables path traversal protection by enforcing that all
-    /// file operations must be within the specified workspace root directory.
-    ///
-    /// # Arguments
-    ///
-    /// * `workspace_roots` - The root directory paths to enforce for security
-    pub fn with_workspace_roots(workspace_roots: Vec<PathBuf>) -> Self {
-        Self { provider: ExecuteCommandProvider::with_workspace_roots(workspace_roots) }
-    }
-
-    /// Execute a command with proper JSON-RPC error handling.
-    ///
-    /// This method converts ExecuteCommandProvider results into proper LSP-compatible
-    /// JSON-RPC responses, mapping errors to appropriate error codes according to
-    /// LSP 3.17 specification.
-    ///
-    /// # Arguments
-    ///
-    /// * `command` - The command name to execute
-    /// * `arguments` - Optional array of command arguments
-    ///
-    /// # Returns
-    ///
-    /// A Result containing either the command result Value or a JsonRpcError
-    /// with appropriate error code and contextual information.
-    ///
-    /// # Error Codes
-    ///
-    /// - `-32602`: Invalid arguments or missing required parameters
-    /// - `-32601`: Unknown or unsupported command
-    /// - `-32603`: Internal errors including security violations
-    pub fn execute(
-        &self,
-        command: &str,
-        arguments: Option<&Vec<Value>>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        // Convert arguments to the format expected by ExecuteCommandProvider
-        let args = arguments.cloned().unwrap_or_default();
-
-        match self.provider.execute_command(command, args) {
-            Ok(result) => Ok(Some(result)),
-            Err(e) => {
-                // Map errors to appropriate JSON-RPC error codes
-                let error_code = if e.contains("Missing") || e.contains("argument") {
-                    -32602 // InvalidParams
-                } else if e.contains("Unknown command") {
-                    -32601 // MethodNotFound
-                } else if e.contains("Path traversal")
-                    || e.contains("security")
-                    || e.contains("workspace root")
-                {
-                    -32603 // InternalError (security)
-                } else {
-                    -32603 // InternalError (general)
-                };
-
-                Err(JsonRpcError {
-                    code: error_code,
-                    message: format!("Execute command failed: {}", e),
-                    data: Some(json!({
-                        "command": command,
-                        "errorType": "executeCommand",
-                        "originalError": e
-                    })),
-                })
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-    use tempfile::tempdir;
-
-    #[test]
-    fn test_supported_commands_includes_run_critic() {
-        let commands = get_supported_commands();
-        assert!(
-            commands.contains(&"perl.runCritic".to_string()),
-            "perl.runCritic should be in supported commands list"
-        );
-    }
-
-    #[test]
-    fn test_execute_command_run_critic_builtin() -> Result<(), Box<dyn std::error::Error>> {
-        let temp_dir = tempdir()?;
-        let temp_file = temp_dir.path().join("test_violations_unit.pl");
-
-        // Create a temporary file with violations
-        let test_content = r#"#!/usr/bin/perl
-# Test file with policy violations
-my $variable = 42;
-print "Value: $variable\n";
-"#;
-
-        fs::write(&temp_file, test_content)?;
-
-        let provider =
-            ExecuteCommandProvider::with_workspace_roots(vec![temp_dir.path().to_path_buf()]);
-        let result = provider.execute_command(
-            "perl.runCritic",
-            vec![Value::String(temp_file.display().to_string())],
-        );
-
-        // Verify result
-        assert!(result.is_ok(), "perl.runCritic command should execute successfully");
-
-        let result_value = result?;
-        assert_eq!(result_value["status"], "success", "Command should succeed");
-        assert!(result_value["violations"].is_array(), "Should return violations array");
-        assert!(
-            result_value["analyzerUsed"].is_string(),
-            "Should indicate which analyzer was used"
-        );
-
-        // Should detect missing 'use strict' and 'use warnings'
-        let violations =
-            result_value["violations"].as_array().ok_or("expected violations array")?;
-        assert!(!violations.is_empty(), "Should detect policy violations");
-
-        // Check for specific violations
-        let has_strict_violation = violations.iter().any(|v| {
-            v["policy"]
-                .as_str()
-                .map(|p| p.contains("RequireUseStrict") || p.contains("strict"))
-                .unwrap_or(false)
-        });
-
-        let has_warnings_violation = violations.iter().any(|v| {
-            v["policy"]
-                .as_str()
-                .map(|p| p.contains("RequireUseWarnings") || p.contains("warnings"))
-                .unwrap_or(false)
-        });
-
-        assert!(has_strict_violation, "Should detect missing 'use strict'");
-        assert!(has_warnings_violation, "Should detect missing 'use warnings'");
-        Ok(())
-    }
-
-    #[test]
-    fn test_execute_command_invalid_command() {
-        let provider = ExecuteCommandProvider::new();
-        let result = provider.execute_command("perl.invalidCommand", vec![]);
-        assert!(result.is_err(), "Invalid command should return error");
-        assert!(result.unwrap_err().contains("Unknown command"), "Should indicate unknown command");
-    }
-
-    #[test]
-    fn test_execute_command_run_critic_missing_file() -> Result<(), Box<dyn std::error::Error>> {
-        let provider = ExecuteCommandProvider::new();
-        let result = provider.execute_command(
-            "perl.runCritic",
-            vec![Value::String("/tmp/nonexistent.pl".to_string())],
-        );
-
-        assert!(result.is_ok(), "Should handle missing files gracefully");
-        let result_value = result?;
-        assert_eq!(result_value["status"], "error", "Should report error status");
-        assert!(
-            result_value["error"]
-                .as_str()
-                .ok_or("expected error string")?
-                .contains("File not found"),
-            "Should indicate file not found"
-        );
-        Ok(())
-    }
-
-    // ============= MUTATION HARDENING TESTS =============
-    // These tests target specific surviving mutants to achieve ≥80% mutation score
-
-    #[test]
-    fn test_command_routing_perl_run_tests() -> Result<(), Box<dyn std::error::Error>> {
-        let temp_dir = tempdir()?;
-        let temp_file = temp_dir.path().join("test_run_tests.pl");
-
-        // Create a test file to ensure we get a specific result
-        let test_content = "#!/usr/bin/perl\nuse strict;\nuse warnings;\nprint 'test';\n";
-        fs::write(&temp_file, test_content)?;
-
-        let provider =
-            ExecuteCommandProvider::with_workspace_roots(vec![temp_dir.path().to_path_buf()]);
-        let result = provider
-            .execute_command("perl.runTests", vec![Value::String(temp_file.display().to_string())]);
-
-        // Verify the command was routed correctly and executed
-        assert!(result.is_ok(), "perl.runTests should execute successfully");
-        let result_value = result?;
-        assert!(result_value.is_object(), "Should return a structured result");
-        assert!(result_value["success"].is_boolean(), "Should have success field");
-        assert!(result_value["output"].is_string(), "Should have output field");
-        Ok(())
-    }
-
-    #[test]
-    fn test_command_routing_perl_run_file() -> Result<(), Box<dyn std::error::Error>> {
-        let temp_dir = tempdir()?;
-        let temp_file = temp_dir.path().join("test_run_file.pl");
-
-        // Create a test file
-        let test_content = "#!/usr/bin/perl\nuse strict;\nuse warnings;\nprint 'hello world';\n";
-        fs::write(&temp_file, test_content)?;
-
-        let provider =
-            ExecuteCommandProvider::with_workspace_roots(vec![temp_dir.path().to_path_buf()]);
-        let result = provider
-            .execute_command("perl.runFile", vec![Value::String(temp_file.display().to_string())]);
-
-        // Verify the command was routed correctly
-        assert!(result.is_ok(), "perl.runFile should execute successfully");
-        let result_value = result?;
-        assert!(result_value.is_object(), "Should return a structured result");
-        assert!(result_value["success"].is_boolean(), "Should have success field");
-        assert!(result_value["output"].is_string(), "Should have output field");
-        Ok(())
-    }
-
-    #[test]
-    fn test_command_routing_perl_run_test_sub() -> Result<(), Box<dyn std::error::Error>> {
-        let temp_dir = tempdir()?;
-        let temp_file = temp_dir.path().join("test_run_test_sub.pl");
-
-        // Create a test file with a subroutine
-        let test_content = "#!/usr/bin/perl\nuse strict;\nuse warnings;\nsub test_sub { print 'test executed'; }\n";
-        fs::write(&temp_file, test_content)?;
-
-        let provider =
-            ExecuteCommandProvider::with_workspace_roots(vec![temp_dir.path().to_path_buf()]);
-        let result = provider.execute_command(
-            "perl.runTestSub",
-            vec![
-                Value::String(temp_file.display().to_string()),
-                Value::String("test_sub".to_string()),
-            ],
-        );
-
-        // Verify the command was routed correctly
-        assert!(result.is_ok(), "perl.runTestSub should execute successfully");
-        let result_value = result?;
-        assert!(result_value.is_object(), "Should return a structured result");
-        assert!(result_value["success"].is_boolean(), "Should have success field");
-        assert!(result_value["subroutine"].is_string(), "Should have subroutine field");
-        Ok(())
-    }
-
-    #[test]
-    fn test_command_routing_perl_debug_tests() -> Result<(), Box<dyn std::error::Error>> {
-        let temp_dir = tempdir()?;
-        let temp_file = temp_dir.path().join("test_debug.pl");
-        fs::write(&temp_file, "print 'debug';")?;
-
-        let provider =
-            ExecuteCommandProvider::with_workspace_roots(vec![temp_dir.path().to_path_buf()]);
-        let result = provider.execute_command(
-            "perl.debugTests",
-            vec![Value::String(temp_file.display().to_string())],
-        );
-
-        // Verify the command was routed correctly
-        assert!(result.is_ok(), "perl.debugTests should execute successfully");
-        let result_value = result?;
-        assert!(result_value.is_object(), "Should return a structured result");
-        assert_eq!(result_value["success"], false, "Debug should indicate not implemented");
-        assert!(result_value["output"].is_string(), "Should have output field");
-        Ok(())
-    }
-
-    #[test]
-    fn test_parameter_validation_missing_file_path() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test with no arguments
-        let result = provider.execute_command("perl.runTests", vec![]);
-        assert!(result.is_err(), "Should fail with missing file path");
-        assert!(result.unwrap_err().contains("Missing file path argument"));
-
-        // Test with null argument
-        let result = provider.execute_command("perl.runTests", vec![Value::Null]);
-        assert!(result.is_err(), "Should fail with null argument");
-        assert!(result.unwrap_err().contains("Missing file path argument"));
-
-        // Test with number instead of string
-        let result = provider
-            .execute_command("perl.runTests", vec![Value::Number(serde_json::Number::from(123))]);
-        assert!(result.is_err(), "Should fail with non-string argument");
-        assert!(result.unwrap_err().contains("Missing file path argument"));
-    }
-
-    #[test]
-    fn test_parameter_validation_missing_subroutine_name() -> Result<(), Box<dyn std::error::Error>>
-    {
-        let temp_dir = tempdir()?;
-        let temp_file = temp_dir.path().join("test_missing_sub.pl");
-        fs::write(&temp_file, "sub test {}")?;
-
-        let provider =
-            ExecuteCommandProvider::with_workspace_roots(vec![temp_dir.path().to_path_buf()]);
-        let file_arg = temp_file.display().to_string();
-
-        // Test runTestSub with only file path, missing subroutine name
-        let result =
-            provider.execute_command("perl.runTestSub", vec![Value::String(file_arg.clone())]);
-
-        assert!(result.is_err(), "Should fail with missing subroutine name");
-        // It might fail with path resolution if file doesn't exist, but here it exists
-        let err = result.err().ok_or("expected error")?;
-        assert!(err.contains("Missing subroutine name argument"));
-
-        // Test with null second argument
-        let result =
-            provider.execute_command("perl.runTestSub", vec![Value::String(file_arg), Value::Null]);
-
-        assert!(result.is_err(), "Should fail with null subroutine name");
-        let err = result.err().ok_or("expected error")?;
-        assert!(err.contains("Missing subroutine name argument"));
-        Ok(())
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn test_normalize_file_path_uri_handling() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test file:// URI scheme stripping
-        let normalized = provider.normalize_file_path("file:///tmp/test.pl");
-        assert_eq!(normalized, "/tmp/test.pl", "Should strip file:// prefix");
-
-        // Test regular path (no URI scheme)
-        let normalized = provider.normalize_file_path("/tmp/test.pl");
-        assert_eq!(normalized, "/tmp/test.pl", "Should leave regular paths unchanged");
-
-        // Test empty string
-        let normalized = provider.normalize_file_path("");
-        assert_eq!(normalized, "", "Should handle empty strings");
-    }
-
-    #[test]
-    fn test_is_test_file_logic() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test .t extension
-        assert!(provider.is_test_file("test.t"), "Should recognize .t files");
-        assert!(provider.is_test_file("path/to/test.t"), "Should recognize .t files in paths");
-
-        // Test /t/ directory
-        assert!(provider.is_test_file("/path/t/test.pl"), "Should recognize files in t/ directory");
-
-        // Test 'test' in name
-        assert!(
-            provider.is_test_file("test_file.pl"),
-            "Should recognize files with 'test' in name"
-        );
-        assert!(provider.is_test_file("my_test.pl"), "Should recognize files with 'test' in name");
-
-        // Test non-test files
-        assert!(!provider.is_test_file("regular.pl"), "Should not recognize regular files");
-        assert!(!provider.is_test_file("module.pm"), "Should not recognize modules");
-    }
-
-    #[test]
-    fn test_format_command_result_structure() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test successful result
-        let output = std::process::Output {
-            status: mock_status(0),
-            stdout: b"test output".to_vec(),
-            stderr: b"".to_vec(),
-        };
-
-        let result = provider.format_command_result(output, None);
-        assert_eq!(result["success"], true, "Should indicate success");
-        assert_eq!(result["output"], "test output", "Should include stdout");
-        assert_eq!(result["error"], Value::Null, "Should have null error for success");
-
-        // Test with extra field
-        let output = std::process::Output {
-            status: mock_status(0),
-            stdout: b"test".to_vec(),
-            stderr: b"".to_vec(),
-        };
-
-        let result = provider
-            .format_command_result(output, Some(("command", Value::String("perl".to_string()))));
-        assert_eq!(result["command"], "perl", "Should include extra field");
-    }
-
-    #[test]
-    fn test_format_command_result_failure() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test failed result
-        let output = std::process::Output {
-            status: mock_status(1),
-            stdout: b"partial output".to_vec(),
-            stderr: b"error message".to_vec(),
-        };
-
-        let result = provider.format_command_result(output, None);
-        assert_eq!(result["success"], false, "Should indicate failure");
-        assert_eq!(result["output"], "partial output", "Should include stdout");
-        assert_eq!(result["error"], "error message", "Should include stderr as error");
-    }
-
-    #[test]
-    fn test_format_violation_structure() {
-        let provider = ExecuteCommandProvider::new();
-
-        let violation = provider.format_violation(
-            "TestPolicy",
-            "Test description",
-            "Test explanation",
-            3,
-            10,
-            5,
-            "/tmp/test.pl",
-        );
-
-        assert_eq!(violation["policy"], "TestPolicy");
-        assert_eq!(violation["description"], "Test description");
-        assert_eq!(violation["explanation"], "Test explanation");
-        assert_eq!(violation["severity"], 3);
-        assert_eq!(violation["line"], 10);
-        assert_eq!(violation["column"], 5);
-        assert_eq!(violation["file"], "/tmp/test.pl");
-    }
-
-    #[test]
-    fn test_format_critic_error_structure() {
-        let provider = ExecuteCommandProvider::new();
-
-        let error_response =
-            provider.format_critic_error("Test error message".to_string(), "test_analyzer");
-
-        assert_eq!(error_response["status"], "error");
-        assert_eq!(error_response["error"], "Test error message");
-        assert!(error_response["violations"].is_array());
-        assert_eq!(error_response["violationCount"], 0);
-        assert_eq!(error_response["analyzerUsed"], "test_analyzer");
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn test_run_critic_file_exists_check() -> Result<(), Box<dyn std::error::Error>> {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test with non-existent file
-        let result = provider.run_critic("/tmp/definitely_nonexistent_file_12345.pl");
-        assert!(result.is_ok(), "Should handle missing files gracefully");
-
-        let result_value = result?;
-        assert_eq!(result_value["status"], "error", "Should report error status");
-        assert!(
-            result_value["error"]
-                .as_str()
-                .ok_or("expected error string")?
-                .contains("File not found"),
-            "Should indicate file not found"
-        );
-        assert_eq!(result_value["analyzerUsed"], "none", "Should indicate no analyzer used");
-        Ok(())
-    }
-
-    #[test]
-    fn test_run_builtin_critic_with_valid_file() -> Result<(), Box<dyn std::error::Error>> {
-        let provider = ExecuteCommandProvider::new();
-
-        // Create a temporary file with content
-        let test_content = "#!/usr/bin/perl\nmy $var = 42;\nprint $var;\n";
-        let temp_file = "/tmp/test_builtin_critic.pl";
-        fs::write(temp_file, test_content)?;
-
-        let path = Path::new(temp_file);
-        let result = provider.run_builtin_critic(path);
-
-        // Clean up
-        fs::remove_file(temp_file).ok();
-
-        assert!(result.is_ok(), "Built-in critic should execute successfully");
-        let result_value = result?;
-        assert_eq!(result_value["status"], "success");
-        assert!(result_value["violations"].is_array());
-        assert_eq!(result_value["analyzerUsed"], "builtin");
-        Ok(())
-    }
-
-    #[test]
-    fn test_command_exists_behavior() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test with a command that definitely exists
-        let exists = provider.command_exists("echo");
-        // Note: We can't assert true here because the mutation test replaces return value
-        // But we can verify it returns a boolean (this always passes but validates function call)
-        assert!(matches!(exists, true | false), "Should return a boolean");
-
-        // Test with a command that definitely doesn't exist
-        let exists = provider.command_exists("definitely_nonexistent_command_12345");
-        // This should be false, but mutation testing may change the logic
-        assert!(matches!(exists, true | false), "Should return a boolean");
-    }
-
-    #[test]
-    fn test_all_command_routing_paths() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test each command path individually to ensure routing logic is tested
-        let commands_to_test = vec![
-            "perl.runTests",
-            "perl.runFile",
-            "perl.runTestSub",
-            "perl.debugTests",
-            "perl.runCritic",
-        ];
-
-        for command in commands_to_test {
-            let args = if command == "perl.runTestSub" {
-                vec![
-                    Value::String("/tmp/test.pl".to_string()),
-                    Value::String("test_sub".to_string()),
-                ]
-            } else {
-                vec![Value::String("/tmp/test.pl".to_string())]
-            };
-
-            let result = provider.execute_command(command, args);
-
-            // Each command should either succeed or fail gracefully
-            // but should never panic or be unhandled
-            match result {
-                Ok(value) => {
-                    assert!(value.is_object(), "Successful results should be objects");
-                }
-                Err(error) => {
-                    // Errors should be meaningful
-                    assert!(!error.is_empty(), "Error messages should not be empty");
-                }
-            }
-        }
-    }
-
-    // ============= ADDITIONAL MUTATION KILLER TESTS =============
-    // These tests specifically target remaining surviving mutants
-
-    #[test]
-    fn test_execute_command_return_value_mutations() -> Result<(), Box<dyn std::error::Error>> {
-        let temp_dir = tempdir()?;
-        let temp_file = temp_dir.path().join("test_mutations.pl");
-        fs::write(&temp_file, "print 'test';")?;
-
-        let provider =
-            ExecuteCommandProvider::with_workspace_roots(vec![temp_dir.path().to_path_buf()]);
-
-        // This test ensures that execute_command cannot return Ok(Default::default())
-        // when it should return meaningful data
-        let result = provider.execute_command(
-            "perl.debugTests",
-            vec![Value::String(temp_file.display().to_string())],
-        );
-
-        assert!(result.is_ok(), "Should return Ok");
-        let result_value = result?;
-
-        // Verify it's not just a default empty object
-        assert!(result_value.is_object(), "Should return an object");
-        assert!(
-            result_value.as_object().ok_or("expected object")?.contains_key("success"),
-            "Should have success field"
-        );
-        assert!(
-            result_value.as_object().ok_or("expected object")?.contains_key("output"),
-            "Should have output field"
-        );
-
-        // The result should be meaningful, not just Default::default()
-        assert_ne!(
-            result_value,
-            Value::Object(serde_json::Map::new()),
-            "Should not be empty object"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_run_tests_logic_operators() -> Result<(), Box<dyn std::error::Error>> {
-        let temp_dir = tempdir()?;
-        let provider = ExecuteCommandProvider::new();
-
-        // Create test files to test is_test_file && command_exists logic
-        let test_file_t = temp_dir.path().join("mutation_test.t");
-        let non_test_file = temp_dir.path().join("mutation_test.pl");
-
-        fs::write(&test_file_t, "use Test::More; ok(1); done_testing();")?;
-        fs::write(&non_test_file, "print 'hello world';")?;
-
-        // Test with .t file (should attempt to use prove if available)
-        let result = provider.run_tests(&test_file_t);
-        assert!(result.is_ok(), "Should handle .t files");
-        let result_value = result?;
-        assert!(result_value["success"].is_boolean(), "Should have boolean success");
-        assert!(result_value["output"].is_string(), "Should have string output");
-
-        // Test with non-test file (should use perl directly)
-        let result = provider.run_tests(&non_test_file);
-        assert!(result.is_ok(), "Should handle .pl files");
-        let result_value = result?;
-        assert!(result_value["success"].is_boolean(), "Should have boolean success");
-        assert!(result_value["output"].is_string(), "Should have string output");
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_is_test_file_operator_mutations() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test various combinations to catch || to && mutations
-
-        // Should be true - ends with .t
-        let result = provider.is_test_file("script.t");
-        assert!(result, "Files ending in .t should be test files");
-
-        // Should be true - contains /t/
-        let result = provider.is_test_file("/path/t/script.pl");
-        assert!(result, "Files in t/ directory should be test files");
-
-        // Should be true - contains 'test'
-        let result = provider.is_test_file("my_test.pl");
-        assert!(result, "Files with 'test' in name should be test files");
-
-        // Should be false - none of the above
-        let result = provider.is_test_file("regular.pl");
-        assert!(!result, "Regular files should not be test files");
-
-        // Edge case - file that would be false if && was used instead of ||
-        let result = provider.is_test_file("test"); // has 'test' but not .t or /t/
-        assert!(result, "Should be true with OR logic");
-    }
-
-    #[test]
-    fn test_run_builtin_critic_arithmetic_mutations() -> Result<(), Box<dyn std::error::Error>> {
-        let provider = ExecuteCommandProvider::new();
-
-        // Create a test file with content at known line/column positions
-        let test_content = "#!/usr/bin/perl\n# Line 2\nmy $var = 42;\nprint $var;\n";
-        let temp_file = "/tmp/test_arithmetic_mutations.pl";
-        fs::write(temp_file, test_content)?;
-
-        let path = Path::new(temp_file);
-        let result = provider.run_builtin_critic(path);
-
-        fs::remove_file(temp_file).ok();
-
-        assert!(result.is_ok(), "Should analyze file successfully");
-        let result_value = result?;
-
-        // Verify that line/column arithmetic is correct
-        let violations =
-            result_value["violations"].as_array().ok_or("expected violations array")?;
-        for violation in violations {
-            let line = violation["line"].as_u64().ok_or("expected line number")?;
-            let column = violation["column"].as_u64().ok_or("expected column number")?;
-
-            // Line and column should be positive (+ 1 conversions work)
-            assert!(line > 0, "Line numbers should be positive (not result of - or * mutations)");
-            assert!(
-                column > 0,
-                "Column numbers should be positive (not result of - or * mutations)"
-            );
-
-            // Should be reasonable values for a short file
-            assert!(line <= 10, "Line numbers should be reasonable");
-            assert!(column <= 100, "Column numbers should be reasonable");
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_format_command_result_negation_mutation() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test successful status - should NOT be negated
-        let success_output = std::process::Output {
-            status: mock_status(0),
-            stdout: b"success".to_vec(),
-            stderr: b"".to_vec(),
-        };
-
-        let result = provider.format_command_result(success_output, None);
-        assert_eq!(result["success"], true, "Success status should not be negated");
-        assert_eq!(result["error"], Value::Null, "Success should have null error");
-
-        // Test failure status - should properly indicate failure
-        let failure_output = std::process::Output {
-            status: mock_status(1),
-            stdout: b"output".to_vec(),
-            stderr: b"error".to_vec(),
-        };
-
-        let result = provider.format_command_result(failure_output, None);
-        assert_eq!(result["success"], false, "Failure status should be false");
-        assert_eq!(result["error"], "error", "Failure should include stderr");
-    }
-
-    /// Verifies that mock_status() correctly round-trips exit codes on both platforms.
-    /// This documents the POSIX (high-byte encoding) vs Windows (direct code) behavior.
-    #[test]
-    fn test_exit_status_roundtrip() {
-        let ok = mock_status(0);
-        assert_eq!(ok.code(), Some(0), "Exit code 0 should round-trip correctly");
-        assert!(ok.success(), "Exit code 0 should be success");
-
-        let fail = mock_status(1);
-        assert_eq!(fail.code(), Some(1), "Exit code 1 should round-trip correctly");
-        assert!(!fail.success(), "Exit code 1 should be failure");
-    }
-
-    #[test]
-    fn test_format_functions_not_default() -> Result<(), Box<dyn std::error::Error>> {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test format_violation doesn't return Default::default()
-        let violation = provider.format_violation(
-            "TestPolicy",
-            "Description",
-            "Explanation",
-            3,
-            5,
-            10,
-            "/tmp/test.pl",
-        );
-
-        assert_ne!(
-            violation,
-            Value::Object(serde_json::Map::new()),
-            "Should not return empty object"
-        );
-        assert!(violation.is_object(), "Should return structured object");
-        assert!(!violation.as_object().ok_or("expected object")?.is_empty(), "Should have content");
-
-        // Test format_critic_error doesn't return Default::default()
-        let error = provider.format_critic_error("Test error".to_string(), "test");
-
-        assert_ne!(error, Value::Object(serde_json::Map::new()), "Should not return empty object");
-        assert!(error.is_object(), "Should return structured object");
-        assert!(!error.as_object().ok_or("expected object")?.is_empty(), "Should have content");
-        Ok(())
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn test_normalize_file_path_not_hardcoded() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test that normalize_file_path returns actual processed values, not hardcoded ones
-        let file_uri = "file:///home/user/test.pl";
-        let result = provider.normalize_file_path(file_uri);
-        assert_eq!(result, "/home/user/test.pl", "Should properly strip file:// prefix");
-        assert_ne!(result, "", "Should not return empty string");
-        assert_ne!(result, "xyzzy", "Should not return hardcoded value");
-
-        let regular_path = "/home/user/test.pl";
-        let result = provider.normalize_file_path(regular_path);
-        assert_eq!(result, regular_path, "Should return input unchanged");
-        assert_ne!(result, "", "Should not return empty string");
-        assert_ne!(result, "xyzzy", "Should not return hardcoded value");
-    }
-
-    #[test]
-    fn test_command_exists_not_hardcoded_true() {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test with a command that definitely doesn't exist
-        // This should return false, not hardcoded true
-        let exists = provider.command_exists("definitely_nonexistent_command_xyz_12345");
-
-        // The mutant that returns hardcoded true would fail this test
-        // Note: We can't always assert false due to environment differences,
-        // but we can verify the function actually runs the check
-        assert!(matches!(exists, true | false), "Should return boolean result");
-
-        // Test multiple times to catch inconsistencies from mutations
-        let exists2 = provider.command_exists("definitely_nonexistent_command_xyz_12345");
-        assert_eq!(exists, exists2, "Should be consistent");
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn test_run_critic_file_existence_logic() -> Result<(), Box<dyn std::error::Error>> {
-        let provider = ExecuteCommandProvider::new();
-
-        // Test the file existence negation - ensure ! is not deleted
-        let result = provider.run_critic("/tmp/absolutely_nonexistent_file_xyz_12345.pl");
-
-        assert!(result.is_ok(), "Should handle gracefully");
-        let result_value = result?;
-        assert_eq!(result_value["status"], "error", "Should detect missing file");
-        assert!(
-            result_value["error"]
-                .as_str()
-                .ok_or("expected error string")?
-                .contains("File not found"),
-            "Should indicate file not found"
-        );
-
-        // If the ! in !path.exists() was deleted, this test would fail
-        // because it would try to process a non-existent file
-        Ok(())
-    }
-
-    #[test]
-    fn test_method_return_values_not_defaults() -> Result<(), Box<dyn std::error::Error>> {
-        let provider = ExecuteCommandProvider::new();
-
-        // Create a real test file
-        let test_content = "#!/usr/bin/perl\nuse strict;\nprint 'hello';\n";
-        let temp_file = "/tmp/test_return_values.pl";
-        fs::write(temp_file, test_content)?;
-
-        // Test run_file doesn't return Ok(Default::default())
-        let result = provider.run_file(Path::new(temp_file));
-        assert!(result.is_ok(), "run_file should succeed");
-        let result_value = result?;
-        assert_ne!(
-            result_value,
-            Value::Object(serde_json::Map::new()),
-            "Should not be empty object"
-        );
-
-        // Test run_tests doesn't return Ok(Default::default())
-        let result = provider.run_tests(Path::new(temp_file));
-        assert!(result.is_ok(), "run_tests should succeed");
-        let result_value = result?;
-        assert_ne!(
-            result_value,
-            Value::Object(serde_json::Map::new()),
-            "Should not be empty object"
-        );
-
-        // Test run_test_sub doesn't return Ok(Default::default())
-        let sub_content = "#!/usr/bin/perl\nuse strict;\nsub test_func { print 'test'; }\n";
-        let sub_file = "/tmp/test_sub_return.pl";
-        fs::write(sub_file, sub_content)?;
-
-        let result = provider.run_test_sub(Path::new(sub_file), "test_func");
-        assert!(result.is_ok(), "run_test_sub should succeed");
-        let result_value = result?;
-        assert_ne!(
-            result_value,
-            Value::Object(serde_json::Map::new()),
-            "Should not be empty object"
-        );
-
-        // Clean up
-        fs::remove_file(temp_file).ok();
-        fs::remove_file(sub_file).ok();
-        Ok(())
-    }
-
-    #[test]
-    fn test_execute_command_workspace_security() -> Result<(), Box<dyn std::error::Error>> {
-        // Create a temporary workspace and a file outside it
-        let workspace_dir = std::env::temp_dir().join("perl_lsp_workspace");
-        let outside_file = std::env::temp_dir().join("perl_lsp_outside.pl");
-
-        fs::create_dir_all(&workspace_dir)?;
-        fs::write(&outside_file, "print 'outside';")?;
-
-        let provider = ExecuteCommandProvider::with_workspace_roots(vec![workspace_dir.clone()]);
-
-        // Try to execute the outside file
-        let result = provider.execute_command(
-            "perl.runFile",
-            vec![Value::String(outside_file.to_string_lossy().to_string())],
-        );
-
-        // Clean up
-        fs::remove_dir_all(&workspace_dir).ok();
-        fs::remove_file(&outside_file).ok();
-
-        // Verify security check
-        assert!(result.is_err(), "Should fail execution outside workspace");
-        let error = result.err().ok_or("expected error")?;
-        assert!(
-            error.contains("Path traversal") || error.contains("outside workspace roots"),
-            "Error should indicate security violation: {}",
-            error
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_execute_command_multi_root_security() -> Result<(), Box<dyn std::error::Error>> {
-        // Create two temporary workspaces and a file outside both
-        let workspace_dir1 = std::env::temp_dir().join("perl_lsp_workspace_1");
-        let workspace_dir2 = std::env::temp_dir().join("perl_lsp_workspace_2");
-        let file1 = workspace_dir1.join("test1.pl");
-        let file2 = workspace_dir2.join("test2.pl");
-        let outside_file = std::env::temp_dir().join("perl_lsp_outside_multi.pl");
-
-        fs::create_dir_all(&workspace_dir1)?;
-        fs::create_dir_all(&workspace_dir2)?;
-        fs::write(&file1, "print 'file1';")?;
-        fs::write(&file2, "print 'file2';")?;
-        fs::write(&outside_file, "print 'outside';")?;
-
-        let provider = ExecuteCommandProvider::with_workspace_roots(vec![
-            workspace_dir1.clone(),
-            workspace_dir2.clone(),
-        ]);
-
-        // 1. Should succeed for file in workspace 1
-        let result1 = provider.execute_command(
-            "perl.runFile",
-            vec![Value::String(file1.to_string_lossy().to_string())],
-        );
-        assert!(result1.is_ok(), "Should allow execution in workspace 1");
-
-        // 2. Should succeed for file in workspace 2
-        let result2 = provider.execute_command(
-            "perl.runFile",
-            vec![Value::String(file2.to_string_lossy().to_string())],
-        );
-        assert!(result2.is_ok(), "Should allow execution in workspace 2");
-
-        // 3. Should fail for outside file
-        let result3 = provider.execute_command(
-            "perl.runFile",
-            vec![Value::String(outside_file.to_string_lossy().to_string())],
-        );
-        assert!(result3.is_err(), "Should fail execution outside both workspaces");
-        let error3 = result3.err().ok_or("expected error")?;
-        assert!(
-            error3.contains("Path traversal") || error3.contains("outside workspace roots"),
-            "Error should indicate security violation: {}",
-            error3
-        );
-
-        // Clean up
-        fs::remove_dir_all(&workspace_dir1).ok();
-        fs::remove_dir_all(&workspace_dir2).ok();
-        fs::remove_file(&outside_file).ok();
-
-        Ok(())
-    }
-}
